Counting Problem
Given a string S, count the number of times each character is present in the string.

Print the character and number of times that character is present in the string separated by a space.

Each character and the number of times it is present should be written in a new line

The order of characters should be in Lexicographical order