N Queens Variant
The n-queens puzzle is the problem of placing n queens on an n×n chessboard such that no two queens attack each other.

The n-queens puzzle is the problem of placing n queens on an n×n chessboard such that no two queens attack each other.

Given an integer n, print the number of distinct solutions to the n-queens puzzle.